+++ src/main/resources/view/HelpPanel.css
@@ -1,0 +1,20 @@
+.help-panel {
+    -fx-background-color: -DP_1_ELEVATION;
+}
+
+#helpMessage {
+    -fx-padding: 20 10;
+    -fx-text-fill: white;
+}
+
+#copyButton {
+    -fx-background-color: -DP_1_ELEVATION;
+    -fx-padding: 20 10;
+    -fx-background-insets: 10 0;
+    -fx-text-fill: white;
+}
+
+#copyButton:pressed, #copyButton:default:hover:pressed {
+    -fx-background-color: -DP_2_ELEVATION;
+    -fx-text-fill: white;
+}
