+++ src/main/resources/view/HelpPanel.fxml
@@ -1,0 +1,29 @@
+<?xml version="1.0" encoding="UTF-8"?>
+
+<?import java.net.URL?>
+
+<?import javafx.geometry.Insets?>
+<?import javafx.scene.control.Button?>
+<?import javafx.scene.control.Label?>
+<?import javafx.scene.layout.HBox?>
+<?import javafx.scene.layout.VBox?>
+
+<VBox xmlns="http://javafx.com/javafx/8" xmlns:fx="http://javafx.com/fxml/1" styleClass="help-panel">
+  <stylesheets>
+    <URL value="@HelpPanel.css" />
+  </stylesheets>
+
+  <HBox alignment="CENTER">
+      <Label fx:id="helpMessage" text="Label">
+        <HBox.margin>
+          <Insets right="5.0"/>
+        </HBox.margin>
+      </Label>
+      <Button fx:id="copyButton" mnemonicParsing="false" onAction="#copyUrl" text="Copy URL">
+        <HBox.margin>
+          <Insets left="5.0"/>
+        </HBox.margin>
+      </Button>
+  </HBox>
+
+</VBox>
